Version 0.8.0
In this version a implementation of the usage of the model, with torch, and the agent for controller_ai, witch basically uses an DQN network to train the model how to land in the moon.
The model is not good, for now the game simulation is way different from the original openai gym lunar lander.