Reduce E_DECAY
To make epsilon be reduced faster.

This helped the model a lot, see the comparison of how epsilon was affected before and after this commit:
- Step 100: before ~= 0.60577, after = 0.95121
- Step 200: before ~= 0.36695, after = 0.90481
- Step 300: before ~= 0.22229, after = 0.86067
- Step 400: before ~= 0.13465, after = 0.81868
- Step 500: before ~= 0.08157, after = 0.77875

The implementation of the lunar lander implementation (by Deeplearning.ai) had good results at 500 steps, while the model before 500 steps was very bad, and the main reason seems to be epsilon, for the previous implementation, in 500 steps, approximately 77% of the actions taken by the agent were completely random, now, also with 500 steps, only 8% of the steps are random, which is a much more plausible amount.