Version 1.0.0
The reinforcement learning model works, and the simulation is sufficiently faithful to the original Lunar Lander made by OpenAI gym, so... Version 1!

After 400 iterations, the lander is stable in the air, using the thrusters well to adjust its own rotation, even with the random inputs, it still tries to spend a little fuel after landing, and at 500 steps it practically stops using fuel after landing, but at 500 steps it takes a long time to land, spending a lot of time in the air just balancing itself.